move tests fix snyk vulnerabilit
