+++ node-runner-cli/requirements.txt
@@ -9,11 +9,11 @@
 -e git+https://github.com/radixdlt/python-core-client@1a61f8a1e2acae859522bbff923b0042ec8955e5#egg=core-client
 bech32==1.2.0
 boto3==1.20.33
-botocore~=1.31.2
-certifi==2023.5.7; python_version >= '3.6'
+botocore==1.23.33
+certifi==2023.7.22 ; python_version >= '3.6'
 cffi==1.15.1
 charset-normalizer==2.0.12; python_version >= '3'
-cryptography~=41.0.0
+cryptography==36.0.1
 deepdiff==5.8.0
 deepmerge==0.3.0
 ecdsa==0.17.0
@@ -30,15 +30,11 @@ pluggy==1.2.0; python_version >= '3.7'
 pycparser==2.21
 pytest==7.4.0
 python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
-pyyaml~=5.4.1
-requests~=2.31.0
+pyyaml==6.0.0
+requests==2.26.0
 s3transfer==0.5.2; python_version >= '3.6'
 six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
 tomli==2.0.1; python_version < '3.11'
 typing-extensions==4.7.1; python_version < '3.8'
 urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
 zipp==3.15.0; python_version >= '3.7'
-
-setuptools~=65.6.3
-urllib3~=1.26.16
-python-dateutil~=2.8.2
