+++ node-runner-cli/version/__init__.py
@@ -1,2 +1,2 @@
-__version__= "2.0.rcnet-v2-rc1-2-gf70af18"
-__base_version__= "test-release-fix"
+__version__= "2.0.rcnet-v2-rc2-79-gac3a031"
+__base_version__= "2.0.rcnet-v2-rc2"
